Reorganize API module
