+++ src/api/mod.rs
@@ -1,21 +1,10 @@
 pub mod fetch;
 pub use fetch::GithubRepositoryName;
 
-use graphql_client::{reqwest::post_graphql, GraphQLQuery};
+pub mod graphql;
+
 use reqwest::{header, Client, ClientBuilder};
 use std::error::Error;
-
-#[allow(clippy::upper_case_acronyms)]
-type URI = String;
-type DateTime = String;
-
-#[derive(GraphQLQuery)]
-#[graphql(
-    schema_path = "src/api/schema.json",
-    query_path = "src/api/test.graphql",
-    response_derives = "Debug"
-)]
-struct TestQuery;
 
 fn get_client_builder() -> ClientBuilder {
     Client::builder()
@@ -38,22 +27,4 @@ pub fn get_github_client() -> Result<Client, Box<dyn Error + Send + Sync>> {
     headers.insert(header::AUTHORIZATION, token_header);
 
     Ok(get_client_builder().default_headers(headers).build()?)
-}
-
-pub async fn query_test() -> Result<(), Box<dyn Error + Send + Sync>> {
-    let client = get_github_client()?;
-
-    let vars = test_query::Variables {};
-
-    // let body = TestQuery::build_query(vars);
-    // let response = client.get("https://api.github.com/graphql").json(&body).send().await?;
-    // let response_body: test_query::ResponseData = response.json().await?;
-
-    let response_body =
-        post_graphql::<TestQuery, _>(&client, "https://api.github.com/graphql", vars).await?;
-
-    // println!("{:#?}", response);
-    println!("{response_body:#?}");
-
-    Ok(())
 }
