+++ src/main.rs
@@ -40,7 +40,7 @@ where
         .unwrap()
         .as_secs();
 
-    api::query_test().await?;
+    api::graphql::graphql_query_test().await?;
 
     let (weights, urls, test_mode) = input::cli::get_inputs(args.into_iter())?;
